move assets db code to its own package
