+++ apps/designer/app/shared/db/index.ts
@@ -5,4 +5,3 @@ export * as misc from "./misc.server";
 export * as user from "./user.server";
 export * from "./canvas.server";
 export * as breakpoints from "./breakpoints.server";
-export * as assets from "./assets.server";
