+++ packages/asset-uploader/src/db/load.ts
@@ -1,0 +1,16 @@
+import { prisma, Project } from "@webstudio-is/prisma-client";
+
+export const loadByProject = async (projectId?: Project["id"]) => {
+  if (typeof projectId !== "string") {
+    throw new Error("Tree ID required");
+  }
+
+  const assets = await prisma.asset.findMany({
+    where: { projectId },
+    orderBy: {
+      createdAt: "desc",
+    },
+  });
+
+  return assets;
+};
