+++ packages/asset-uploader/src/index.ts
@@ -1,59 +1,2 @@
-import path from "path";
-import {
-  unstable_createFileUploadHandler,
-  unstable_parseMultipartFormData,
-} from "@remix-run/node";
-import { s3UploadHandler } from "./targets/s3/handler";
-import { uploadToS3 } from "./targets/s3/uploader";
-import { uploadToDisk } from "./targets/disk/upload";
-import { assetEnvVariables, fsEnvVariables, s3EnvVariables } from "./schema";
-
-const isS3Upload = s3EnvVariables.safeParse(process.env).success;
-const fsUploadVars = fsEnvVariables.parse(process.env);
-const commonUploadVars = assetEnvVariables.parse(process.env);
-
-// user inputs the max value in mb and we transform it to bytes
-export const MAX_UPLOAD_SIZE = parseInt(commonUploadVars.MAX_UPLOAD_SIZE) * 1e6;
-
-export const uploadAssets = async ({
-  request,
-  projectId,
-  db,
-  dirname,
-}: {
-  request: Request;
-  projectId: string;
-  db: unknown;
-  dirname: string;
-}) => {
-  const uploads = path.join(dirname, "../public");
-  const directory = path.join(uploads, fsUploadVars.FILE_UPLOAD_PATH);
-  const formData = await unstable_parseMultipartFormData(
-    request,
-    isS3Upload
-      ? (file) =>
-          s3UploadHandler({
-            file,
-            maxPartSize: MAX_UPLOAD_SIZE,
-          })
-      : unstable_createFileUploadHandler({
-          maxPartSize: MAX_UPLOAD_SIZE,
-          directory,
-          file: ({ filename }) => filename,
-        })
-  );
-  if (isS3Upload) {
-    return await uploadToS3({
-      projectId,
-      formData,
-      db,
-    });
-  } else {
-    return await uploadToDisk({
-      projectId,
-      formData,
-      folderInPublic: fsUploadVars.FILE_UPLOAD_PATH,
-      db,
-    });
-  }
-};
+export * from "./upload";
+export * from "./db/index.server";
