+++ packages/asset-uploader/src/targets/disk/upload.ts
@@ -1,18 +1,16 @@
 import path from "path";
 import sharp from "sharp";
+import { create } from "../../";
 import { ImagesUpload } from "../../schema";
 
 export const uploadToDisk = async ({
   formData,
   projectId,
   folderInPublic,
-  db,
 }: {
   formData: FormData;
   projectId: string;
   folderInPublic: string;
-  // @todo Stop passing db here
-  db: any;
 }) => {
   const imagesInfo = ImagesUpload.parse(formData.getAll("image"));
   const allInfo = imagesInfo.map(async (image) => {
@@ -28,7 +26,7 @@ export const uploadToDisk = async ({
       metadata,
     };
 
-    const newAsset = await db.assets.create(projectId, data);
+    const newAsset = await create(projectId, data);
 
     return newAsset;
   });
